+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Add
 - Add test option for DictZip CLI.
 - Add DictZipInputStream consructor that get a filename as argument.
+- More tests.
 
 ### Changed
 - README: there is no test dependency for commons-io.