+++ CHANGELOG.md
@@ -8,9 +8,9 @@ All notable changes to this project will be documented in this file.
 - More tests.
 
 ### Changed
-- README: there is no test dependency for commons-io.
 - Default buffer size in the DictZipInputStream was changed to 8192 byte
   from 512 byte as same as java.io.BufferedInputStream.
+- Test static utility isFileBinaryEquals becomes an external library.
 
 ### Fixed
 - Check return value of InputStream.skip() method in the utility method